+++ src/Facades/Soap.php
@@ -6,11 +6,11 @@ namespace RicorocksDigitalAgency\Soap\Facades;
 
 use Illuminate\Support\Facades\Facade;
 use RicorocksDigitalAgency\Soap\Contracts\Request;
-use RicorocksDigitalAgency\Soap\Header;
-use RicorocksDigitalAgency\Soap\HeaderSet;
-use RicorocksDigitalAgency\Soap\Inclusion;
-use RicorocksDigitalAgency\Soap\OptionSet;
 use RicorocksDigitalAgency\Soap\Parameters\Node;
+use RicorocksDigitalAgency\Soap\Support\Header;
+use RicorocksDigitalAgency\Soap\Support\Scopes\HeaderSet;
+use RicorocksDigitalAgency\Soap\Support\Scopes\Inclusion;
+use RicorocksDigitalAgency\Soap\Support\Scopes\OptionSet;
 use SoapVar;
 
 /**
