+++ src/Support/Scopes/Scoped.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace RicorocksDigitalAgency\Soap\Support;
+namespace RicorocksDigitalAgency\Soap\Support\Scopes;
 
 abstract class Scoped
 {
